Update Materials.java
The Rocks still lack Unique color but at least we have them organized now. i still have yet to finish all the minerals i created for these rocks but for now a lot of it is done. However ive found quite a lot of rocks and minor semi precious gems that haven't been implemented into GT6 next time i have a chance im going to implement most of the rest of the gems base besides the tigers eye, seriously that needs 2 varieties not 7